onyx/shared/pyproject.toml
[tool.poetry]
name = "onyx-shared"
version = "0.1.0"
description = ""
authors = ["Tay Nguyen <15603942+hotay@users.noreply.github.com>"]
readme = "README.md"
packages = [{ include = "onyx", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = { version = "^2.0.30", optional = true }
orjson = "^3.9.10"
structlog = "^23.3.0"
"delegator.py" = "^0.1.1"
sentry-sdk = { extras = ["sqlalchemy"], version = "^2.0.1" }
slack-sdk = "3.31.0"
punq = "^0.7.0"

[tool.poetry.extras]
orm = ["sqlalchemy"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
